[Token] Move the error into token namspace
The variants contained inside the error type are all somehow related to
the conversion from or into the internal tokenized state.

Like the Token struct itself this is not exactly a feature of a state
tracker but actually a common element of decoder, encoder and the state
tracker.
