+++ src/token.rs
@@ -1,3 +1,5 @@
+use failure::Fail;
+
 /// A raw bencode token
 #[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Debug)]
 pub enum Token<'a> {
@@ -11,4 +13,37 @@ pub enum Token<'a> {
     Num(&'a str),
     /// The end of a list or dictionary
     End,
+}
+
+/// An encoding or decoding error
+#[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Debug, Fail)]
+pub enum Error {
+    #[fail(display = "Saw the wrong type of token: {}", _0)]
+    /// Wrong type of token detected.
+    InvalidState(String),
+    #[fail(display = "Keys were not sorted")]
+    /// Keys were not sorted.
+    UnsortedKeys,
+    #[fail(display = "Reached EOF in the middle of a message")]
+    /// EOF reached to early.
+    UnexpectedEof,
+    #[fail(display = "Malformed number of unexpected character: {}", _0)]
+    /// Unexpected characters detected.
+    SyntaxError(String),
+    #[fail(display = "Maximum nesting depth exceeded")]
+    /// Exceeded the recursion limit.
+    NestingTooDeep,
+}
+
+impl Error {
+    pub fn unexpected(expected: &str, got: char, offset: usize) -> Self {
+        Error::SyntaxError(format!(
+            "Expected {}, got {:?} at offset {}",
+            expected, got, offset
+        ))
+    }
+
+    pub fn invalid_state(expected: &str) -> Self {
+        Error::InvalidState(expected.to_owned())
+    }
 }
